Prevent creating empty directories
When files are synced from the build directory to the zip directory, if
the include/exclude rules would have resulted in an empty directory,
that empty directory would be created and included in the ZIP.

By just skipping over any directories in the iterator, we prevent that
from happening.

If the directory needs to be created, it is being created just before
the file is written anyway.